# Conflicts:
#	build.xml
#	src/haven/Buff.java
#	src/haven/Defer.java
#	src/haven/Fightsess.java
#	src/haven/GItem.java
#	src/haven/GameUI.java
#	src/haven/Glob.java
#	src/haven/Gob.java
#	src/haven/GobHealth.java
#	src/haven/HavenPanel.java
#	src/haven/KinInfo.java
#	src/haven/Listbox.java
#	src/haven/MainFrame.java
#	src/haven/MapView.java
#	src/haven/MenuGrid.java
#	src/haven/OCache.java
#	src/haven/OptWnd.java
#	src/haven/RenderList.java
#	src/haven/ResDrawable.java
#	src/haven/RootWidget.java
#	src/haven/Session.java
#	src/haven/UI.java
#	src/haven/Widget.java

+++ src/haven/Widget.java
@@ -354,7 +354,7 @@ public class Widget {
 		    st.push(((Widget)st.pop()).sz);
 		} else if(op == 'w') {
 		    synchronized(ui) {
-			st.push(ui.widgets.get((Integer)st.pop()));
+			st.push(ui.getwidget((Integer)st.pop()));
 		    }
 		} else if(op == 'x') {
 		    st.push(((Coord)st.pop()).x);
@@ -483,35 +483,50 @@ public class Widget {
 	}
 	return(false);
     }
-	
+
     public void gotfocus() {
 	if(focusctl && (focused != null)) {
 	    focused.hasfocus = true;
 	    focused.gotfocus();
 	}
     }
-	
+
+    public void dispose() {
+    }
+
+    public void rdispose() {
+	for(Widget ch = child; ch != null; ch = ch.next)
+	    ch.rdispose();
+	dispose();
+    }
+
+    public void remove() {
+	subscriptions.forEach(Subscription::unsubscribe);
+	subscriptions.clear();
+	if(canfocus)
+	    setcanfocus(false);
+	if(parent != null) {
+	    unlink();
+	    parent.cdestroy(this);
+	}
+	ui.removed(this);
+    }
+
     public void reqdestroy() {
 	destroy();
     }
 
     public void destroy() {
-	subscriptions.forEach(Subscription::unsubscribe);
-	subscriptions.clear();
-	if(canfocus)
-	    setcanfocus(false);
-	unlink();
-	parent.cdestroy(this);
+	remove();
+	rdispose();
     }
-    
+
+    /* XXX: Should be renamed to cremove at this point. */
     public void cdestroy(Widget w) {
     }
-	
+
     public int wdgid() {
-	Integer id = ui.rwidgets.get(this);
-	if(id == null)
-	    return(-1);
-	return(id);
+	return(ui.widgetid(this));
     }
 
     public void lostfocus() {
@@ -520,7 +535,7 @@ public class Widget {
 	    focused.lostfocus();
 	}
     }
-	
+
     public void setfocus(Widget w) {
 	if(focusctl) {
 	    if(w != focused) {
@@ -538,8 +553,11 @@ public class Widget {
 		    last.hasfocus = false;
 		    last.lostfocus();
 		}
-		if((ui != null) && ui.rwidgets.containsKey(w) && ui.rwidgets.containsKey(this))
-		    wdgmsg("focus", ui.rwidgets.get(w));
+		if((ui != null) && (w != null) && (wdgid() >= 0)) {
+		    int id = w.wdgid();
+		    if(id >= 0)
+			wdgmsg("focus", id);
+		}
 	    }
 	    if((parent != null) && visible && canfocus)
 		parent.setfocus(this);
@@ -547,7 +565,7 @@ public class Widget {
 	    parent.setfocus(w);
 	}
     }
-	
+
     public void setcanfocus(boolean canfocus) {
 	this.autofocus = this.canfocus = canfocus;
 	if(parent != null) {
@@ -558,7 +576,7 @@ public class Widget {
 	    }
 	}
     }
-	
+
     public void newfocusable(Widget w) {
 	if(focusctl) {
 	    if(focused == null)
@@ -568,7 +586,7 @@ public class Widget {
 		parent.newfocusable(w);
 	}
     }
-	
+
     public void delfocusable(Widget w) {
 	if(focusctl) {
 	    if((focused != null) && focused.hasparent(w)) {
@@ -579,7 +597,7 @@ public class Widget {
 		parent.delfocusable(w);
 	}
     }
-	
+
     private void findfocus() {
 	/* XXX: Might need to check subwidgets recursively */
 	focused = null;
@@ -622,7 +640,7 @@ public class Widget {
 	    if(tid < 0) {
 		setfocus(null);
 	    } else {
-		Widget w = ui.widgets.get(tid);
+		Widget w = ui.getwidget(tid);
 		if(w != null) {
 		    if(w.canfocus)
 			setfocus(w);
